refactor(common): move common tests in own directory
