+++ packages/common/src/__tests__/utils.spec.js
@@ -1,4 +1,4 @@
-const util = require('./utils');
+const util = require('../utils');
 
 describe('Utils test', () => {
 	describe('#promiseTimeout', () => {
